Cleaning worktrees should use worktree list
