+++ src/tests/test_helpers.rs
@@ -1,36 +1,10 @@
-use std::{env::current_dir, path::PathBuf, str::Split};
+use std::{env::current_dir, path::PathBuf};
 
 use crate::{
     tests::test_setup::{setup, teardown},
-    utils::{get_all_branch_names, get_current_branch_name},
-    Context, RepoType, commands::git_command,
+    utils::{get_all_branch_names, get_current_branch_name, get_all_worktree_names},
+    Context, RepoType
 };
-
-fn clean_worktree_name(worktree: &String) -> Option<String> {
-    let mut parts: Split<char> = worktree.as_str().split('[');
-
-    if parts.clone().count() < 2 {
-        return None;
-    }
-
-    let name = parts
-        .next_back()
-        .expect("Couldn't get second item")
-        .strip_suffix(']')
-        .expect("Couldn't strip suffix")
-        .to_string();
-
-    Some(name)
-}
-
-pub fn get_all_worktree_names(repo_path: &PathBuf) -> Vec<String> {
-    git_command(vec!["worktree", "list"], PathBuf::from(repo_path))
-        .expect("Couldn't get worktree names")
-        .output
-        .iter()
-        .filter_map(clean_worktree_name)
-        .collect::<Vec<String>>()
-}
 
 pub fn assert_branches(context: Context, branches: Vec<String>) {
     assert_eq!(get_all_branch_names(&context.repo_path), branches);
@@ -45,7 +19,9 @@ pub fn assert_branch_does_not_exist(context: Context, branch: String) {
 }
 
 pub fn assert_worktree_does_not_exist(context: Context, worktree: String) {
-    assert!(!get_all_worktree_names(&context.repo_path).contains(&worktree));
+    let worktrees = get_all_worktree_names(&context.repo_path).expect("Couldn't get list of worktrees");
+
+    assert!(!worktrees.contains(&worktree));
 }
 
 pub fn assert_current_branch(context: &Context, branch: String) {
