+++ src/utils.rs
@@ -1,4 +1,4 @@
-use std::{env, path::PathBuf};
+use std::{env, path::PathBuf, str::Split};
 
 use crate::{commands::git_command, CommandWorkingDirectory};
 
@@ -25,6 +25,34 @@ pub fn get_all_branch_names(repo_path: &PathBuf) -> Vec<String> {
         .iter()
         .map(clean_branch_name)
         .collect::<Vec<String>>()
+}
+
+fn clean_worktree_name(worktree: &String) -> Option<String> {
+    let mut parts: Split<char> = worktree.as_str().split('[');
+
+    if parts.clone().count() < 2 {
+        return None;
+    }
+
+    let name = parts
+        .next_back()
+        .expect("Couldn't get second item")
+        .strip_suffix(']')
+        .expect("Couldn't strip suffix")
+        .to_string();
+
+    Some(name)
+}
+
+pub fn get_all_worktree_names(repo_path: &PathBuf) -> Result<Vec<String>, String> {
+    let names = git_command(vec!["worktree", "list"], PathBuf::from(repo_path))
+        .expect("Couldn't get worktree names")
+        .output
+        .iter()
+        .filter_map(clean_worktree_name)
+        .collect::<Vec<String>>();
+
+    Ok(names)
 }
 
 pub fn get_main_branch_name(repo_path: &PathBuf) -> String {
